The beginnings of UI animation
Every shape will have an animation which is a list of timed states(keyframes).
Every frame of the game, each playing animation will be advanced 1 frame,
which is interpolating the state values between the current frame and the next keyframe.

Currently this is only a class implementation that is on its own. Next will
be to hook these into the game and content pipeline, along with adding
convenience functions for playback and such.